chore: create home page
